Handle dead unix socket (input source)
During 'targetctl power off/on' the qemu unix socket
is being deleted and then created again.

The ser2tcp does not disconnect when unix socket dies.

In this commit, we fix that, when unix socket (input source)
receives empty data then it raises a socket.error which
cause to disconnect and try to re-connect when unix socket
is available again.

Thus, during 'targetctl power off/on', ser2tcp behaves
as expected.